feat: 비즈니스 로직 테스트를 통과하는 성별 수정 로직 구현

+++ was/member/src/main/java/org/example/application/MemberService.kt
@@ -5,6 +5,7 @@ import org.example.domain.entity.Member
 import org.example.domain.enums.Gender
 import org.example.domain.enums.Gender.*
 import org.example.domain.repository.MemberRepository
+import org.example.exception.exceptions.GenderNotValidException
 import org.example.exception.exceptions.MemberNotFoundException
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -27,12 +28,17 @@ open class MemberService(private val memberRepository: MemberRepository) {
 
     @Transactional
     open fun deleteMember(member: Member) {
-        memberRepository.findById(member.id).orElseThrow{MemberNotFoundException()}
-        memberRepository.delete(member)
+        if(memberRepository.existsById(member.id)) {
+            memberRepository.delete(member)
+            return
+        }
+
+        throw MemberNotFoundException()
     }
 
     private fun findGenderByInput(genderString: String): Gender {
         if(genderString == MALE.value) return MALE;
-        return FEMALE;
+        if(genderString == FEMALE.value) return FEMALE;
+        throw GenderNotValidException()
     }
 }
